Set 'selected' attr via Mojo controller
... instead of setting the 'selected' attribute directly.  The reason
for this commit is because how the 'selected' attribute was handled was
changed in Mojolicious version 6.16 (see [the release
changes](https://metacpan.org/changes/release/SRI/Mojolicious-7.74#L784)
for more info) and the tests for this plugin had begun failing as of
this Mojo version.  The behaviour of the 'selected' attribute was also
changed in Mojolicious version 5.73 (it's necessary to diff the 5.73 <->
5.72 versions in order to see this; it's not mentioned in the Changes)
which meant that setting the `selected` attribute to a true value no
longer uses `selected='selected'` but simply the presence of the
`selected` attribute is sufficient to set it to a true value.  This
commit should fix the current failures occurring on CPAN Testers.